Rename extractimages to extractimageurls
