+++ development/pkg/extractimageurls/kubernetes.go
@@ -1,4 +1,4 @@
-package extractimages
+package extractimageurls
 
 import (
 	"errors"
