+++ development/pkg/extractimageurls/prowjob.go
@@ -1,4 +1,4 @@
-package extractimages
+package extractimageurls
 
 import (
 	"k8s.io/test-infra/prow/config"
