+++ development/pkg/extractimageurls/prowjob_test.go
@@ -1,4 +1,4 @@
-package extractimages
+package extractimageurls
 
 import (
 	"reflect"
